+++ RELEASE_NOTES.md
@@ -6,10 +6,12 @@
 - `Removed` be.baur.sds.common.Date and DateTime.
 - `Changed` DataType.getContentType() to getType().
 - `Changed` Interval.from(String, Class) to from(String, Function).
+- `Changed` AbstractStringType to a generic type.
 - Added Schema.registerDataType(), and get/isDataType().
 - Added abstract Type, which AnyType and NodeType extend.
 - Added static valueOf() methods for native RangedTypes.
 - Added GMonthDay custom type example and tests.
+- Added IBAN custom type example and tests.
 
 ## Compatibility